requirements.txt
# Arxiv Paper Collector - Dependencies

# Arxiv API
arxiv>=1.4.0

# YAML configuration
PyYAML>=6.0

# Template engine
Jinja2>=3.1.0

# Date handling
python-dateutil>=2.8.0

# Logging
colorlog>=6.7.0

# PDF compilation (LaTeX must be installed separately)
# No additional Python packages needed for PDF compilation

